+++ packages/ui/useScroll/useScroll.js
@@ -1147,10 +1147,13 @@ const useScrollBase = (props) => {
 				stop();
 			}
 		} else { // scrollMode 'native'
+			const roundedTargetX = scrollContentHandle.current.scrollPos.left < targetX ? Math.ceil(targetX) : Math.floor(targetX);
+			const roundedTargetY = scrollContentHandle.current.scrollPos.top < targetY ? Math.ceil(targetY) : Math.floor(targetY);
+
 			if (animate) {
-				scrollContentHandle.current.scrollToPosition(targetX, targetY, 'smooth');
+				scrollContentHandle.current.scrollToPosition(roundedTargetX, roundedTargetY, 'smooth');
 			} else {
-				scrollContentHandle.current.scrollToPosition(targetX, targetY, 'instant');
+				scrollContentHandle.current.scrollToPosition(roundedTargetX, roundedTargetY, 'instant');
 			}
 
 			if (props.start) {