+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact project, newest changes 
 ### Fixed
 
 - `ui/Scroller` to calculate correctly targetX and targetY when scrolling to a position
+- `spotlight` navigation from an element to a different container
 
 ## [5.3.0] - 2025-09-24