+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The following is a curated list of changes in the Enact project, newest changes on the top.
 
+## [5.3.1] - 2025-10-14
+
+### Fixed
+
+- `ui/Scroller` to calculate correctly targetX and targetY when scrolling to a position
+- `spotlight` navigation from an element to a different container
+
 ## [5.3.0] - 2025-09-24
 
 ### Added